+++ kamu-adapter-graphql/src/scalars/metadata.rs
@@ -7,30 +7,37 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
+use crate::queries::Account;
 use crate::scalars::*;
 
 use async_graphql::*;
 use chrono::{DateTime, Utc};
 
 /////////////////////////////////////////////////////////////////////////////////////////
-// MetadataBlockHashed
+// MetadataBlockExtended
 /////////////////////////////////////////////////////////////////////////////////////////
 
 #[derive(SimpleObject, Clone, PartialEq, Eq)]
-pub struct MetadataBlockHashed {
+pub struct MetadataBlockExtended {
     pub block_hash: Multihash,
-    pub system_time: DateTime<Utc>,
     pub prev_block_hash: Option<Multihash>,
+    pub system_time: DateTime<Utc>,
+    pub author: Account,
     pub event: MetadataEvent,
 }
 
-impl MetadataBlockHashed {
-    pub fn new<H: Into<Multihash>, B: Into<MetadataBlock>>(block_hash: H, block: B) -> Self {
+impl MetadataBlockExtended {
+    pub fn new<H: Into<Multihash>, B: Into<MetadataBlock>>(
+        block_hash: H,
+        block: B,
+        author: Account,
+    ) -> Self {
         let b = block.into();
         Self {
             block_hash: block_hash.into(),
             prev_block_hash: b.prev_block_hash,
             system_time: b.system_time,
+            author,
             event: b.event,
         }
     }